+++ src/main/java/com/example/ifclubserver/member/domain/dto/request/CreateMemberRequest.java
@@ -1,0 +1,13 @@
+package com.example.ifclubserver.member.domain.dto.request;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class CreateMemberRequest {
+    private String name;
+    private String phone;
+    private String email;
+
+}
