+++ src/main/java/com/example/ifclubserver/member/domain/dto/request/MemberUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.example.ifclubserver.member.domain.dto;
+package com.example.ifclubserver.member.domain.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,7 +7,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
-public class MemberUpdateForm {
+public class MemberUpdateRequest {
     private String name;
     private String phone;
     private String email;
